pyproject.toml
[tool.poetry]
name = "popcat"
version = "0.1.0"
description = ""
authors = ["Muhammad Umair <umairmuqeem7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
python-telegram-bot = "^21.6"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
